+++ opensc-opendnie/trunk/src/tools/util.c
@@ -136,9 +136,9 @@ void util_print_binary(FILE *f, const u8 *buf, int count)
 	(void) fflush(f);
 }
 
-void util_hex_dump(FILE *f, const u8 *in, int len, const char *sep)
+void util_hex_dump(FILE *f, const u8 *in, size_t len, const char *sep)
 {
-	int i;
+	size_t i;
 
 	for (i = 0; i < len; i++) {
 		if (sep != NULL && i)
